Move order chat form into separate file for testing
